requirements.txt
flask==3.1.2
cryptography==46.0.3
psycopg2-binary==2.9.11
cachetools==6.2.1
flask_limiter==4.0.0
Flask-Admin==1.6.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.1.0
python-dotenv==1.2.1
requests==2.32.5
redis==7.0.1
gunicorn==23.0.0
gevent>=1.4
setuptools<81